+++ CHANGELOG.md
@@ -1,9 +1,10 @@
 # Changelog
 
-All notable changes to this project will be documented in this file.
-
 ## [Unreleased]
-- 
+- Per-axis grid snap (UI + snapping)
+- Array tools skeleton (linear, radial, grid)
+- Align pack: center, match rotation/size
+- Command palette skeleton + toolbar button
 
 ## [0.1.0] - 2026-06-20
 ### Added